Move CrawlerProcess to scrapy/utils/command.py
